Test_wsfexv1: Agrego uso de variables de ser necesario

+++ tests/test_wsfexv1.py
@@ -23,17 +23,19 @@ from pyafipws.wsaa import WSAA
 from pyafipws.wsfexv1 import WSFEXv1
 
 # Para utilizar variables de entorno en Travis 
-cert = os.environ['CERT']
-pkey = os.environ['PKEY']
 
-CERT = cert.replace(r'\n', '\n')
-PKEY = pkey.replace(r'\n', '\n')
+if not os.path.exists('rei.crt'):
+    cert = os.environ['CERT']
+    pkey = os.environ['PKEY']
 
-with open('rei.crt', 'w', encoding='utf-8') as f:
-    f.write(CERT)
+    CERT = cert.replace(r'\n', '\n')
+    PKEY = pkey.replace(r'\n', '\n')
+
+    with open('rei.crt', 'w', encoding='utf-8') as f:
+        f.write(CERT)    
     
-with open('rei.key', 'w', encoding='utf-8') as f:
-    f.write(PKEY)
+    with open('rei.key', 'w', encoding='utf-8') as f:
+        f.write(PKEY)
 
 WSDL = "https://wswhomo.afip.gov.ar/wsfexv1/service.asmx?WSDL"
 CUIT = os.environ['CUIT']
